:hammer: shift NamedContent to plugin interface
