+++ pyexcel_io/plugin_api/__init__.py
@@ -1,3 +1,3 @@
-from .abstract_sheet import ISheet, ISheetWriter  # noqa: F401
+from .abstract_sheet import ISheet, ISheetWriter, NamedContent  # noqa: F401
 from .abstract_reader import IReader  # noqa: F401
 from .abstract_writer import IWriter  # noqa: F401
